SimScen updates
